미들웨어 두개 하나로 합침
