+++ .gitignore
@@ -289,3 +289,5 @@ pyrightconfig.json
 
 .DS_Store
 # End of https://www.toptal.com/developers/gitignore/api/pycharm,python
+logs/
+log/
